Spray Charge + Minor changes
Spray have a very short charge duration now to procrastinate dealing with further code injection.
This is an elegant solution, as it also stops from instant spraying after consuming a potion.
In addition to that, there should also be a minimum charge requirement in order to avoid accidental instant casting of projectile / self spells.